% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{finches}
\alias{finches}
\title{Finch occurrence data from Sanderson (2000)}
\format{
A data frame with 13 observations on 17 variables.
}
\source{
Sanderson (2000)
}
\usage{
data(finches)
}
\description{
Occurrence data for 13 species from 17 sites. Columns are sites, rows are
species. 1's are presences and 0's are absences.
}
\details{
This data set is unchanged from the \strong{cooccur} package (version
1.3), currently archived on CRAN.
}
\examples{
data(finches)
}
\references{
Sanderson JG (2000) "Testing Ecological Patterns: A Well-known Algorithm from
Computer Science Aids the Evaulation of Species Distributions". \emph{American
Scientist} 88(4), 332--339. \url{https://www.jstor.org/stable/27858056}
}
\keyword{datasets}
