% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ca-mjca.r
\name{methods-ca-mjca}
\alias{methods-ca-mjca}
\alias{as_tbl_ord.mjca}
\alias{recover_rows.mjca}
\alias{recover_cols.mjca}
\alias{recover_inertia.mjca}
\alias{recover_conference.mjca}
\alias{recover_coord.mjca}
\alias{recover_aug_rows.mjca}
\alias{recover_aug_cols.mjca}
\alias{recover_aug_coord.mjca}
\title{Functionality for multiple and joint correspondence analysis ('mjca')
objects}
\usage{
\method{as_tbl_ord}{mjca}(x)

\method{recover_rows}{mjca}(x)

\method{recover_cols}{mjca}(x)

\method{recover_inertia}{mjca}(x)

\method{recover_conference}{mjca}(x)

\method{recover_coord}{mjca}(x)

\method{recover_aug_rows}{mjca}(x)

\method{recover_aug_cols}{mjca}(x)

\method{recover_aug_coord}{mjca}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\value{
The recovery generics \verb{recover_*()} return \link[ordr:recoverers]{core model components}, \link[ordr:conference]{distribution of inertia},
\link[ordr:supplementation]{supplementary elements}, and \link[ordr:augmentation]{intrinsic metadata}; but they require methods for each model class to
tell them what these components are.

The generic \code{\link[=as_tbl_ord]{as_tbl_ord()}} returns its input wrapped in the 'tbl_ord'
class. Its methods determine what model classes it is allowed to wrap. It
then provides 'tbl_ord' methods with access to the recoverers and hence to
the model components.
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"mjca"} from the \strong{\link[ca:mjca]{ca}} package.
}
\section{{Single, multiple, and joint CA}}{


Greenacre (2007) proposes joint correspondence analysis (JCA) as a
counterpart to classical multiple correspondence analysis (MCA). These,
together with classical correspondence analysis (CA), are implemented in the
\strong{ca} package.
}

\examples{
# table of admissions and rejections from UC Berkeley
class(UCBAdmissions)
head(as.data.frame(UCBAdmissions))

if (require(ca)) {# {ca}

# perform multiple correspondence analysis
UCBAdmissions \%>\%
  ca::mjca() \%>\%
  as_tbl_ord() \%>\%
  # augment profiles with names, masses, distances, and inertias
  augment_ord() \%>\%
  print() -> admissions_mca

# recover row and column profiles
head(get_rows(admissions_mca))
get_cols(admissions_mca)

# column-standard biplot of factor levels
admissions_mca \%>\%
  ggbiplot() +
  theme_bw() + theme_biplot() +
  geom_origin() +
  # geom_rows_count() +
  geom_cols_point(aes(shape = factor, size = mass)) +
  geom_cols_text_repel(aes(label = level), show.legend = FALSE) +
  scale_size_area(guide = "none") +
  labs(shape = "Factor level")

# column-principal biplot of factor levels
admissions_mca \%>\%
  confer_inertia("colprincipal") \%>\%
  ggbiplot() +
  theme_bw() + theme_biplot() +
  geom_origin() +
  # geom_rows_point(stat = "sum") +
  geom_cols_point(aes(shape = factor, size = mass)) +
  geom_cols_text_repel(aes(label = level), show.legend = FALSE) +
  scale_size_area(guide = "none") +
  labs(shape = "Factor level")

}# {ca}
}
\references{
Greenacre MJ (2007) \emph{Correspondence Analysis in Practice}, Second Edition.
Boca Raton: Chapman & Hall/CRC, ISBN 978-1-58488-616-7.
\url{http://www.carme-n.org/?sec=books2}
}
\seealso{
Other methods for singular value decomposition-based techniques: 
\code{\link{methods-ade4}},
\code{\link{methods-ca-ca}},
\code{\link{methods-candisc-cancor}},
\code{\link{methods-clpca}},
\code{\link{methods-lpca}},
\code{\link{methods-lsvd}},
\code{\link{methods-nipals-empca}},
\code{\link{methods-nipals-nipals}},
\code{\link{methods-pma-cca}},
\code{\link{methods-pma-spc}}

Other models from the \strong{ca} package: 
\code{\link{methods-ca-ca}}
}
\concept{methods for singular value decomposition-based techniques}
\concept{models from the \strong{ca} package}
