% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-psych-fa.R
\name{methods-fa}
\alias{methods-fa}
\alias{as_tbl_ord.fa}
\alias{recover_rows.fa}
\alias{recover_cols.fa}
\alias{recover_inertia.fa}
\alias{recover_coord.fa}
\alias{recover_conference.fa}
\alias{recover_supp_rows.fa}
\alias{recover_supp_cols.fa}
\alias{recover_aug_rows.fa}
\alias{recover_aug_cols.fa}
\alias{recover_aug_coord.fa}
\title{Functionality for factor analysis ('fa') objects}
\usage{
\method{as_tbl_ord}{fa}(x)

\method{recover_rows}{fa}(x)

\method{recover_cols}{fa}(x)

\method{recover_inertia}{fa}(x)

\method{recover_coord}{fa}(x)

\method{recover_conference}{fa}(x)

\method{recover_supp_rows}{fa}(x)

\method{recover_supp_cols}{fa}(x)

\method{recover_aug_rows}{fa}(x)

\method{recover_aug_cols}{fa}(x)

\method{recover_aug_coord}{fa}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\value{
The recovery generics \verb{recover_*()} return \link[ordr:recoverers]{core model components}, \link[ordr:conference]{distribution of inertia},
\link[ordr:supplementation]{supplementary elements}, and \link[ordr:augmentation]{intrinsic metadata}; but they require methods for each model class to
tell them what these components are.

The generic \code{\link[=as_tbl_ord]{as_tbl_ord()}} returns its input wrapped in the 'tbl_ord'
class. Its methods determine what model classes it is allowed to wrap. It
then provides 'tbl_ord' methods with access to the recoverers and hence to
the model components.
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"fa"} as returned by \code{\link[psych:fa]{psych::fa()}}.
}
\details{
Factor analysis of a data matrix relies on an an eigendecomposition of its
correlation matrix, whose eigenvectors (up to weighting) comprise the
variable loadings. For this reason, both column and supplementary row
recoverers retrieve the loadings and inertia is evenly distributed between
them. When computed and returned by \code{\link[psych:fa]{psych::fa()}}, the case scores are
accessible as supplementary elements. Redistribution of inertia commutes
through both score calculations.
}
\examples{
# data frame of Anderson iris species measurements
class(iris)
head(iris)

# fa on iris
iris |>
  ordinate(~psych::fa(., nfactors = 2L, rotate = "varimax", scores = 
                        "regression", fm = "ml"),
           cols = 1:4, augment = c(Sepal.Width, Species)) |>
  print() -> iris_fa

# recover loadings
get_cols(iris_fa, elements = "active")
# recover pseudoinverse of weights
get_cols(iris_fa, elements = "pinv_weight")
# recover scores
head(get_rows(iris_fa, elements = "score"))

# augment column loadings with uniquenesses
(iris_fa <- augment_ord(iris_fa))

# biplot with both loadings and pseudoinverse weights
ggbiplot(iris_fa, scale_cols = 3) +
  geom_rows_point(elements = "score", aes(color = Species)) +
  geom_cols_vector(linetype = "solid") +
  geom_cols_text_radiate(label = names(iris[1:4]), elements = "active", 
                         linetype = "solid") +
  geom_cols_vector(linetype = "dotted", elements = "pinv_weight") +
  geom_cols_text_repel(label = names(iris[1:4]), elements = "pinv_weight", 
                       linetype = "dotted", size = 3.5) +
  # include vector labels in plot window (specific to graphics device)
  scale_x_continuous(limits = ~ range(.x, c(-2, 5))) +
  scale_y_continuous(limits = ~ range(.x, c(-4, 3)))
}
\seealso{
Other methods for eigen-decomposition-based techniques: 
\code{\link{methods-candisc-candisc}},
\code{\link{methods-principal}}

Other models from the psych package: 
\code{\link{methods-principal}}
}
\author{
John Gracey
}
\concept{methods for eigen-decomposition-based techniques}
\concept{models from the psych package}
