% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-pma-spc.r
\name{methods-pma-spc}
\alias{methods-pma-spc}
\alias{as_tbl_ord.SPC}
\alias{recover_coord.SPC}
\alias{recover_rows.SPC}
\alias{recover_cols.SPC}
\alias{recover_inertia.SPC}
\alias{recover_conference.SPC}
\alias{recover_aug_rows.SPC}
\alias{recover_aug_cols.SPC}
\alias{recover_aug_coord.SPC}
\title{Functionality for penalized multivariate analysis ('SPC') objects}
\usage{
\method{as_tbl_ord}{SPC}(x)

\method{recover_coord}{SPC}(x)

\method{recover_rows}{SPC}(x)

\method{recover_cols}{SPC}(x)

\method{recover_inertia}{SPC}(x)

\method{recover_conference}{SPC}(x)

\method{recover_aug_rows}{SPC}(x)

\method{recover_aug_cols}{SPC}(x)

\method{recover_aug_coord}{SPC}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\value{
The recovery generics \verb{recover_*()} return \link[ordr:recoverers]{core model components}, \link[ordr:conference]{distribution of inertia},
\link[ordr:supplementation]{supplementary elements}, and \link[ordr:augmentation]{intrinsic metadata}; but they require methods for each model class to
tell them what these components are.

The generic \code{\link[=as_tbl_ord]{as_tbl_ord()}} returns its input wrapped in the 'tbl_ord'
class. Its methods determine what model classes it is allowed to wrap. It
then provides 'tbl_ord' methods with access to the recoverers and hence to
the model components.
}
\description{
These methods extract data from, and attribute new data to,
objects of class 'SPC' from the \strong{\link[PMA:PMA-package]{PMA}} package.
}
\section{{Penalized matrix decomposition}}{


Witten, Tibshirani, and Hastie (2009) provide a theoretical basis and
computational algorithm for penalized matrix decomposition that specializes
to sparse PCA and to sparse CCA. Their R package \strong{PMA} implements these
specializations as well as one to sparse multiple CCA.
}

\examples{
# data frame of Anderson iris species measurements
class(iris)
head(iris)

if (require(PMA)) {# {PMA}

# compute principal components of scaled measurements
iris_spca <- PMA::SPC(
  as.matrix(iris[, -5L]),
  sumabsv = sqrt(2),
  K = 2L,
  cnames = names(iris)[seq(4L)]
)

# wrap as a 'tbl_ord' object
(iris_spca <- as_tbl_ord(iris_spca))
# bind species classification to observation coordinates
(iris_spca <- mutate_rows(iris_spca, species = iris$Species))

# recover observation principal coordinates and measurement standard coordinates
head(get_rows(iris_spca))
get_cols(iris_spca)
# recover inertia in each dimension
get_inertia(iris_spca)

# augment measurements with names and scaling parameters
(iris_spca <- augment_ord(iris_spca))
# row-principal biplot
iris_spca \%>\%
  augment_ord() \%>\%
  ggbiplot(axis.type = "predictive") +
  theme_bw() + theme_biplot() +
  scale_color_brewer(type = "qual", palette = 2) +
  # -+- text elements are extremely far removed -+-
  geom_cols_axis(aes(label = name, center = meanx)) +
  geom_rows_point(aes(color = species), alpha = .5) +
  ggtitle("Row-principal sparse PCA biplot of Anderson iris measurements") +
  expand_limits(x = c(-.12, .2))

}# {PMA}
}
\references{
Witten DM, Tibshirani R, & Hastie T (2009) "A penalized matrix decomposition,
with applications to sparse principal components and canonical correlation
analysis". \emph{Biostatistics} \strong{10}(3): 515--534.
\doi{10.1093/biostatistics/kxp008}
}
\seealso{
Other methods for singular value decomposition-based techniques: 
\code{\link{methods-ade4}},
\code{\link{methods-ca-ca}},
\code{\link{methods-ca-mjca}},
\code{\link{methods-candisc-cancor}},
\code{\link{methods-clpca}},
\code{\link{methods-lpca}},
\code{\link{methods-lsvd}},
\code{\link{methods-nipals-empca}},
\code{\link{methods-nipals-nipals}},
\code{\link{methods-pma-cca}}

Other models from the \strong{PMA} package: 
\code{\link{methods-pma-cca}}
}
\concept{methods for singular value decomposition-based techniques}
\concept{models from the \strong{PMA} package}
