% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-psych-principal.R
\name{methods-principal}
\alias{methods-principal}
\alias{as_tbl_ord.principal}
\alias{recover_rows.principal}
\alias{recover_cols.principal}
\alias{recover_inertia.principal}
\alias{recover_coord.principal}
\alias{recover_conference.principal}
\alias{recover_supp_rows.principal}
\alias{recover_aug_rows.principal}
\alias{recover_aug_cols.principal}
\alias{recover_aug_coord.principal}
\title{Functionality for principal components analysis ('principal') objects}
\usage{
\method{as_tbl_ord}{principal}(x)

\method{recover_rows}{principal}(x)

\method{recover_cols}{principal}(x)

\method{recover_inertia}{principal}(x)

\method{recover_coord}{principal}(x)

\method{recover_conference}{principal}(x)

\method{recover_supp_rows}{principal}(x)

\method{recover_aug_rows}{principal}(x)

\method{recover_aug_cols}{principal}(x)

\method{recover_aug_coord}{principal}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\value{
The recovery generics \verb{recover_*()} return \link[ordr:recoverers]{core model components}, \link[ordr:conference]{distribution of inertia},
\link[ordr:supplementation]{supplementary elements}, and \link[ordr:augmentation]{intrinsic metadata}; but they require methods for each model class to
tell them what these components are.

The generic \code{\link[=as_tbl_ord]{as_tbl_ord()}} returns its input wrapped in the 'tbl_ord'
class. Its methods determine what model classes it is allowed to wrap. It
then provides 'tbl_ord' methods with access to the recoverers and hence to
the model components.
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"principal"} as returned by \code{\link[psych:principal]{psych::principal()}}.
}
\details{
Principal components analysis (PCA), as performed by \code{\link[psych:principal]{psych::principal()}},
relies on an eigenvalue decomposition (EVD) of the covariance matrix
\eqn{X^TX} of a data set \eqn{X}. \code{\link[psych:principal]{psych::principal()}} returns the right
singular vectors with full inertia \eqn{VD} from the EVD as the loadings
\verb{$loadings}. The scores \verb{$scores} are obtained as \eqn{XV} and are accessible
as supplementary elements.
}
\examples{
# data frame of Anderson iris species measurements
class(iris)
head(iris)

if (require(psych)) {# {psych}

# compute unscaled row-principal components of scaled measurements
iris[, -5] |>
  psych::principal(nfactors = 4, rotate = "none") |>
  as_tbl_ord() |>
  print() -> iris_pca

# recover observation principal coordinates and measurement standard coordinates
head(get_rows(iris_pca))
get_cols(iris_pca)

# augment measurement coordinates with names and scaling parameters
(iris_pca <- augment_ord(iris_pca))

}# {psych}
}
\seealso{
Other methods for eigen-decomposition-based techniques: 
\code{\link{methods-candisc-candisc}},
\code{\link{methods-fa}}

Other models from the psych package: 
\code{\link{methods-fa}}
}
\author{
John Gracey
}
\concept{methods for eigen-decomposition-based techniques}
\concept{models from the psych package}
