% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ca-ca.r
\name{methods-ca-ca}
\alias{methods-ca-ca}
\alias{as_tbl_ord.ca}
\alias{recover_rows.ca}
\alias{recover_cols.ca}
\alias{recover_inertia.ca}
\alias{recover_conference.ca}
\alias{recover_coord.ca}
\alias{recover_aug_rows.ca}
\alias{recover_aug_cols.ca}
\alias{recover_aug_coord.ca}
\title{Functionality for correspondence analysis ('ca') objects}
\usage{
\method{as_tbl_ord}{ca}(x)

\method{recover_rows}{ca}(x)

\method{recover_cols}{ca}(x)

\method{recover_inertia}{ca}(x)

\method{recover_conference}{ca}(x)

\method{recover_coord}{ca}(x)

\method{recover_aug_rows}{ca}(x)

\method{recover_aug_cols}{ca}(x)

\method{recover_aug_coord}{ca}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\value{
The recovery generics \verb{recover_*()} return \link[=recoverers]{core model components}, \link[=conference]{distribution of inertia},
\link[=supplementation]{supplementary elements}, and \link[=augmentation]{intrinsic metadata}; but they require methods for each model class to
tell them what these components are.

The generic \code{\link[=as_tbl_ord]{as_tbl_ord()}} returns its input wrapped in the 'tbl_ord'
class. Its methods determine what model classes it is allowed to wrap. It
then provides 'tbl_ord' methods with access to the recoverers and hence to
the model components.
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"ca"} from the \strong{\link[ca:ca]{ca}} package.
}
\examples{
# table of hair and eye color data collapsed by sex
class(HairEyeColor)
haireye <- as.data.frame(rowSums(HairEyeColor, dims = 2L))
print(haireye)

# use correspondence analysis to construct row and column profiles
haireye \%>\%
  ca::ca() \%>\%
  as_tbl_ord() \%>\%
  print() -> haireye_ca

# recover row and column profiles
get_rows(haireye_ca)
get_cols(haireye_ca)
# augment profiles with names, masses, distances, and inertias
(haireye_ca <- augment_ord(haireye_ca))
}
\seealso{
\link[ordr:methods-correspondence]{ordr::methods-correspondence}

Other methods for singular value decomposition-based techniques: 
\code{\link{methods-ade4}},
\code{\link{methods-ca-mjca}},
\code{\link{methods-candisc-cancor}},
\code{\link{methods-lpca}},
\code{\link{methods-nipals}},
\code{\link{methods-pma-cca}},
\code{\link{methods-pma-spc}}

Other models from the \strong{ca} package: 
\code{\link{methods-ca-mjca}}
}
\concept{methods for singular value decomposition-based techniques}
\concept{models from the \strong{ca} package}
