% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-candisc-cancor.r
\name{methods-candisc-cancor}
\alias{methods-candisc-cancor}
\alias{as_tbl_ord.cancor}
\alias{recover_rows.cancor}
\alias{recover_cols.cancor}
\alias{recover_inertia.cancor}
\alias{recover_coord.cancor}
\alias{recover_conference.cancor}
\alias{recover_supp_rows.cancor}
\alias{recover_supp_cols.cancor}
\alias{recover_aug_rows.cancor}
\alias{recover_aug_cols.cancor}
\alias{recover_aug_coord.cancor}
\title{Functionality for canonical correlation ('cancor') objects}
\usage{
\method{as_tbl_ord}{cancor}(x)

\method{recover_rows}{cancor}(x)

\method{recover_cols}{cancor}(x)

\method{recover_inertia}{cancor}(x)

\method{recover_coord}{cancor}(x)

\method{recover_conference}{cancor}(x)

\method{recover_supp_rows}{cancor}(x)

\method{recover_supp_cols}{cancor}(x)

\method{recover_aug_rows}{cancor}(x)

\method{recover_aug_cols}{cancor}(x)

\method{recover_aug_coord}{cancor}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\value{
The recovery generics \verb{recover_*()} return \link[ordr:recoverers]{core model components}, \link[ordr:conference]{distribution of inertia},
\link[ordr:supplementation]{supplementary elements}, and \link[ordr:augmentation]{intrinsic metadata}; but they require methods for each model class to
tell them what these components are.

The generic \code{\link[=as_tbl_ord]{as_tbl_ord()}} returns its input wrapped in the 'tbl_ord'
class. Its methods determine what model classes it is allowed to wrap. It
then provides 'tbl_ord' methods with access to the recoverers and hence to
the model components.
}
\description{
These methods extract data from, and attribute new data to,
objects of class 'cancor' from the \strong{\link[candisc:candisc-package]{candisc}}
package.
}
\details{
ter Braak (1990) recommends two families of biplots for the canonical
correspondence analysis of data matrices \eqn{X} and \eqn{Y}.

For the first, using structural correlations, either the interset
correlations of \eqn{X} (equivalent to principal coordinates) are plotted
with the intraset correlations of \eqn{Y} (standard coordinates) or
vice-versa, so that their product recovers the inner product matrix
\eqn{XY'}. ter Braak's biplots can then be recovered by balancing the inertia
across the two factors.

For the second, if the variables are distinguished as predictors and
criteria, then the superposition of the interset correlations of the criteria
with the (inertia-free) canonical coefficients of the predictors yields a
biplot that approximates the matrix of coefficients in the multivariate
regression model.

The methods for \code{\link[candisc:cancor]{candisc::cancor()}} mirror those for \code{\link[ordr:wrap-ord]{ordr::cancor_ord()}} in
\strong{ordr}, though the canonical coefficients (hence the canonical scores) are
inversely scaled by \eqn{n - 1}.
}
\examples{
if (require(candisc)) {# {candisc}

# data frame of life-cycle savings across countries
class(LifeCycleSavings)
head(LifeCycleSavings)

# canonical correlation analysis of age distributions and financial factors
savings_cancor <- candisc::cancor(
  LifeCycleSavings[, c("pop15", "pop75")],
  LifeCycleSavings[, c("sr", "dpi", "ddpi")]
)

# wrap as a 'tbl_ord' object
(savings_cancor <- as_tbl_ord(savings_cancor))
# recover canonical weights
get_rows(savings_cancor, elements = "active")
get_cols(savings_cancor, elements = "active")
head(get_rows(savings_cancor, elements = "supplementary"))
head(get_cols(savings_cancor, elements = "supplementary"))
# augment canonical weights with row names and centers
(savings_cancor <- augment_ord(savings_cancor))

# column-standard biplot of structure correlations
# (intraset correlations of first set, interset correlations with second set)
savings_cancor \%>\%
  confer_inertia("columns") \%>\%
  ggbiplot(aes(label = name)) +
  geom_unit_circle() +
  geom_rows_vector(elements = "structure") +
  geom_rows_text_radiate(elements = "structure") +
  geom_cols_point(elements = "structure", color = "forestgreen") +
  geom_cols_text_repel(elements = "structure", color = "forestgreen") +
  expand_limits(x = c(-1, 1), y = c(-1, 1))

}# {candisc}
}
\references{
ter Braak CJF (1990) "Interpreting canonical correlation analysis through
biplots of structure correlations and weights". \emph{Psychometrika} 55(3),
519--531. \doi{10.1007/BF02294765}
}
\seealso{
\link[ordr:methods-cancor]{ordr::methods-cancor}

Other methods for singular value decomposition-based techniques: 
\code{\link{methods-ade4}},
\code{\link{methods-ca-ca}},
\code{\link{methods-ca-mjca}},
\code{\link{methods-lpca}},
\code{\link{methods-nipals-empca}},
\code{\link{methods-nipals-nipals}},
\code{\link{methods-pma-cca}},
\code{\link{methods-pma-spc}}

Other models from the \strong{candisc} package: 
\code{\link{methods-candisc-candisc}}
}
\concept{methods for singular value decomposition-based techniques}
\concept{models from the \strong{candisc} package}
