% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-candisc-candisc.r
\name{methods-candisc-candisc}
\alias{methods-candisc-candisc}
\alias{as_tbl_ord.candisc}
\alias{recover_rows.candisc}
\alias{recover_cols.candisc}
\alias{recover_inertia.candisc}
\alias{recover_coord.candisc}
\alias{recover_conference.candisc}
\alias{recover_supp_rows.candisc}
\alias{recover_supp_cols.candisc}
\alias{recover_aug_rows.candisc}
\alias{recover_aug_cols.candisc}
\alias{recover_aug_coord.candisc}
\title{Functionality for canonical discriminant ('candisc') objects}
\usage{
\method{as_tbl_ord}{candisc}(x)

\method{recover_rows}{candisc}(x)

\method{recover_cols}{candisc}(x)

\method{recover_inertia}{candisc}(x)

\method{recover_coord}{candisc}(x)

\method{recover_conference}{candisc}(x)

\method{recover_supp_rows}{candisc}(x)

\method{recover_supp_cols}{candisc}(x)

\method{recover_aug_rows}{candisc}(x)

\method{recover_aug_cols}{candisc}(x)

\method{recover_aug_coord}{candisc}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\value{
The recovery generics \verb{recover_*()} return \link[ordr:recoverers]{core model components}, \link[ordr:conference]{distribution of inertia},
\link[ordr:supplementation]{supplementary elements}, and \link[ordr:augmentation]{intrinsic metadata}; but they require methods for each model class to
tell them what these components are.

The generic \code{\link[=as_tbl_ord]{as_tbl_ord()}} returns its input wrapped in the 'tbl_ord'
class. Its methods determine what model classes it is allowed to wrap. It
then provides 'tbl_ord' methods with access to the recoverers and hence to
the model components.
}
\description{
These methods extract data from, and attribute new data to,
objects of class 'candisc' from the \strong{\link[candisc:candisc-package]{candisc}}
package.
}
\details{
As implemented in \strong{\link[candisc:candisc-package]{candisc}}, canonical
discriminant analysis arrives at linear discriminant analysis by way of
multivariate linear regression; see documentation in that package for more
detail.

The methods for \code{\link[candisc:candisc]{candisc::candisc()}} agree with those for \code{\link[MASS:lda]{MASS::lda()}} and
extend them to include canonical structure coefficients as supplementary
column points. These are the correlations between the observed and canonical
variables and therefore, like the canonical coefficients recovered as active
column elements, carry no inertia in the model object.

The \code{plot()} method for \code{candisc} generates a biplot of markers for the
centroid and case scores and vectors for the structure coefficients. This can
be reproduced with \code{\link[ordr:ggbiplot]{ordr::ggbiplot()}} by specifying the supplementary
elements (see the examples). This plot is distinct from those possible using
\code{\link[ordr:lda-ord]{ordr::lda_ord()}}: The vectors represent the locations of the observed
variables in canonical space.
}
\examples{
if (require(candisc)) {# {candisc}

# data frame of Anderson iris species measurements
class(iris)
head(iris)

# multivariate linear model (MANOVA)
iris_mlm <-lm(
  cbind(Petal.Length, Sepal.Length, Petal.Width, Sepal.Width) ~ Species,
  data = iris
)
aov(iris_mlm)

# canonical discriminant analysis
( iris_cda <- augment_ord(as_tbl_ord(candisc::candisc(iris_mlm))) )
# biplot of canonical score markers, canonical mean markers, and canonical
# structure coefficient vectors
# `candisc:::plot.candisc()`
plot(
  iris_cda,
  conf = .99, ellipse = TRUE, ellipse.prob = 2/3,
  main = "Canonical discriminant analysis of Anderson's iris data"
)

# recover centroid coordinates and discriminant coefficients
get_rows(iris_cda, elements = "active")
head(get_rows(iris_cda, elements = "score"))
get_cols(iris_cda, elements = "active")
get_cols(iris_cda, elements = "structure")

# monoplot of canonical score markers and canonical mean markers
# (note that ellipses are inequivalent)
iris_monoplot <- ggbiplot(iris_cda, sec.axes = "cols", scale.factor = 3) +
  theme_scaffold() +
  coord_scaffold() +
  stat_rows_ellipse(elements = "score", aes(color = name), level = 2/3) +
  geom_rows_point(aes(color = name, shape = name), size = 5, alpha = .75) +
  geom_rows_point(
    elements = "score",
    aes(color = name, shape = name), alpha = .5
  )
# biplot with raw canonical coefficient vectors
iris_monoplot + geom_cols_vector(aes(label = name))
# biplot with canonical structure coefficient vectors
iris_monoplot +
  geom_unit_circle(scale.factor = 3) +
  geom_cols_vector(elements = "structure", aes(label = name))

}# {candisc}
}
\seealso{
\code{\link[ordr:methods-cancor]{ordr::methods-cancor}}

Other methods for eigen-decomposition-based techniques: 
\code{\link{methods-fa}},
\code{\link{methods-principal}}

Other models from the \strong{candisc} package: 
\code{\link{methods-candisc-cancor}}
}
\concept{methods for eigen-decomposition-based techniques}
\concept{models from the \strong{candisc} package}
