% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-candisc-candisc.r
\name{methods-candisc-candisc}
\alias{methods-candisc-candisc}
\alias{as_tbl_ord.candisc}
\alias{recover_rows.candisc}
\alias{recover_cols.candisc}
\alias{recover_inertia.candisc}
\alias{recover_coord.candisc}
\alias{recover_conference.candisc}
\alias{recover_supp_rows.candisc}
\alias{recover_supp_cols.candisc}
\alias{recover_aug_rows.candisc}
\alias{recover_aug_cols.candisc}
\alias{recover_aug_coord.candisc}
\title{Functionality for canonical discriminant ('candisc') objects}
\usage{
\method{as_tbl_ord}{candisc}(x)

\method{recover_rows}{candisc}(x)

\method{recover_cols}{candisc}(x)

\method{recover_inertia}{candisc}(x)

\method{recover_coord}{candisc}(x)

\method{recover_conference}{candisc}(x)

\method{recover_supp_rows}{candisc}(x)

\method{recover_supp_cols}{candisc}(x)

\method{recover_aug_rows}{candisc}(x)

\method{recover_aug_cols}{candisc}(x)

\method{recover_aug_coord}{candisc}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\value{
The recovery generics \verb{recover_*()} return \link[=recoverers]{core model components}, \link[=conference]{distribution of inertia},
\link[=supplementation]{supplementary elements}, and \link[=augmentation]{intrinsic metadata}; but they require methods for each model class to
tell them what these components are.

The generic \code{\link[=as_tbl_ord]{as_tbl_ord()}} returns its input wrapped in the 'tbl_ord'
class. Its methods determine what model classes it is allowed to wrap. It
then provides 'tbl_ord' methods with access to the recoverers and hence to
the model components.
}
\description{
These methods extract data from, and attribute new data to,
objects of class 'candisc' from the \strong{\link[candisc:candisc-package]{candisc}}
package.
}
\seealso{
\link[ordr:methods-cancor]{ordr::methods-cancor}

Other models from the \strong{candisc} package: 
\code{\link{methods-candisc-cancor}}
}
\concept{methods for eigen-decomposition-based techniques}
\concept{models from the \strong{candisc} package}
