% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-psych-fa.R
\name{methods-fa}
\alias{methods-fa}
\alias{as_tbl_ord.fa}
\alias{recover_rows.fa}
\alias{recover_cols.fa}
\alias{recover_inertia.fa}
\alias{recover_coord.fa}
\alias{recover_conference.fa}
\alias{recover_supp_rows.fa}
\alias{recover_supp_cols.fa}
\alias{recover_aug_rows.fa}
\alias{recover_aug_cols.fa}
\alias{recover_aug_coord.fa}
\title{Functionality for factor analysis ('fa') objects}
\usage{
\method{as_tbl_ord}{fa}(x)

\method{recover_rows}{fa}(x)

\method{recover_cols}{fa}(x)

\method{recover_inertia}{fa}(x)

\method{recover_coord}{fa}(x)

\method{recover_conference}{fa}(x)

\method{recover_supp_rows}{fa}(x)

\method{recover_supp_cols}{fa}(x)

\method{recover_aug_rows}{fa}(x)

\method{recover_aug_cols}{fa}(x)

\method{recover_aug_coord}{fa}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\value{
The recovery generics \verb{recover_*()} return \link[=recoverers]{core model components}, \link[=conference]{distribution of inertia},
\link[=supplementation]{supplementary elements}, and \link[=augmentation]{intrinsic metadata}; but they require methods for each model class to
tell them what these components are.

The generic \code{\link[=as_tbl_ord]{as_tbl_ord()}} returns its input wrapped in the 'tbl_ord'
class. Its methods determine what model classes it is allowed to wrap. It
then provides 'tbl_ord' methods with access to the recoverers and hence to
the model components.
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"fa"} as returned by \code{\link[psych:fa]{psych::fa()}}.
}
\details{
Factor analysis of a data matrix relies on an an eigendecomposition of its
correlation matrix, whose eigenvectors (up to weighting) comprise the
variable loadings. For this reason, both column and supplementary row
recoverers retrieve the loadings and inertia is evenly distributed between
them. When computed and returned by \code{\link[psych:fa]{psych::fa()}}, the case scores are
accessible as supplementary elements. Redistribution of inertia commutes
through both score calculations.
}
\examples{
# data frame of Swiss fertility and socioeconomic indicators
class(swiss)
head(swiss)
# perform factor analysis
swiss_fa <- psych::fa(
  r = swiss, nfactors = 2L, rotate = "varimax", scores = "regression", 
  fm = "ml"
)

# wrap as a 'tbl_ord' object
(swiss_fa <- as_tbl_ord(swiss_fa))

# recover loadings
get_cols(swiss_fa, elements = "active")
# recover scores
head(get_rows(swiss_fa, elements = "score"))

# augment column loadings with uniquenesses
(swiss_fa <- augment_ord(swiss_fa))
}
\seealso{
Other methods for eigen-decomposition-based techniques: 
\code{\link{methods-principal}}

Other models from the psych package: 
\code{\link{methods-principal}}
}
\author{
John Gracey
}
\concept{methods for eigen-decomposition-based techniques}
\concept{models from the psych package}
