% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-logisticpca-lsvd.r,
%   R/methods-logisticpca-lpca.r, R/methods-logisticpca-clpca.r
\name{methods-lpca}
\alias{methods-lpca}
\alias{as_tbl_ord.lsvd}
\alias{recover_rows.lsvd}
\alias{recover_cols.lsvd}
\alias{recover_coord.lsvd}
\alias{augmentation_rows.lsvd}
\alias{augmentation_cols.lsvd}
\alias{augmentation_coord.lsvd}
\alias{as_tbl_ord.lpca}
\alias{recover_rows.lpca}
\alias{recover_cols.lpca}
\alias{recover_coord.lpca}
\alias{augmentation_rows.lpca}
\alias{augmentation_cols.lpca}
\alias{augmentation_coord.lpca}
\alias{as_tbl_ord.clpca}
\alias{recover_rows.clpca}
\alias{recover_cols.clpca}
\alias{recover_coord.clpca}
\alias{augmentation_rows.clpca}
\alias{augmentation_cols.clpca}
\alias{augmentation_coord.clpca}
\title{Functionality for logistic PCA and logistic SVD objects}
\usage{
\method{as_tbl_ord}{lsvd}(x)

\method{recover_rows}{lsvd}(x)

\method{recover_cols}{lsvd}(x)

\method{recover_coord}{lsvd}(x)

\method{augmentation_rows}{lsvd}(x)

\method{augmentation_cols}{lsvd}(x)

\method{augmentation_coord}{lsvd}(x)

\method{as_tbl_ord}{lpca}(x)

\method{recover_rows}{lpca}(x)

\method{recover_cols}{lpca}(x)

\method{recover_coord}{lpca}(x)

\method{augmentation_rows}{lpca}(x)

\method{augmentation_cols}{lpca}(x)

\method{augmentation_coord}{lpca}(x)

\method{as_tbl_ord}{clpca}(x)

\method{recover_rows}{clpca}(x)

\method{recover_cols}{clpca}(x)

\method{recover_coord}{clpca}(x)

\method{augmentation_rows}{clpca}(x)

\method{augmentation_cols}{clpca}(x)

\method{augmentation_coord}{clpca}(x)
}
\arguments{
\item{x}{An ordination object.}
}
\description{
These methods extract data from, and attribute new data to,
objects of class \code{"lpca"} and \code{"lsvd"} from the
\strong{\link[logisticPCA:logisticPCA-package]{logisticPCA}} package.

These methods extract data from, and attribute new data to,
objects of class \code{"lpca"} and \code{"lsvd"} from the
\strong{\link[logisticPCA:logisticPCA-package]{logisticPCA}} package.

These methods extract data from, and attribute new data to,
objects of class \code{"lpca"} and \code{"lsvd"} from the
\strong{\link[logisticPCA:logisticPCA-package]{logisticPCA}} package.
}
\section{{Logistic SVD-based models}}{


Landgraf and Lee (2020) provide dimension reduction tools for binary data in
their package \strong{logisticPCA}. The signature functions
\code{\link[logisticPCA:logisticPCA]{logisticPCA::logisticPCA()}}, \code{\link[logisticPCA:logisticSVD]{logisticPCA::logisticSVD()}}, and
\code{\link[logisticPCA:convexLogisticPCA]{logisticPCA::convexLogisticPCA()}} have \verb{*_ord()} wrappers that add row and
column names from the input matrix to the output matrices.



Landgraf and Lee (2020) provide dimension reduction tools for binary data in
their package \strong{logisticPCA}. The signature functions
\code{\link[logisticPCA:logisticPCA]{logisticPCA::logisticPCA()}}, \code{\link[logisticPCA:logisticSVD]{logisticPCA::logisticSVD()}}, and
\code{\link[logisticPCA:convexLogisticPCA]{logisticPCA::convexLogisticPCA()}} have \verb{*_ord()} wrappers that add row and
column names from the input matrix to the output matrices.



Landgraf and Lee (2020) provide dimension reduction tools for binary data in
their package \strong{logisticPCA}. The signature functions
\code{\link[logisticPCA:logisticPCA]{logisticPCA::logisticPCA()}}, \code{\link[logisticPCA:logisticSVD]{logisticPCA::logisticSVD()}}, and
\code{\link[logisticPCA:convexLogisticPCA]{logisticPCA::convexLogisticPCA()}} have \verb{*_ord()} wrappers that add row and
column names from the input matrix to the output matrices.
}

\examples{
# site-species data frame of Sanderson Galapagos finches data
data(finches, package = "cooccur")
class(finches)
finches[seq(6L), seq(6L)]

# logistic singular value decomposition
finches \%>\%
  t() \%>\%
  logisticSVD_ord() \%>\%
  as_tbl_ord() \%>\%
  print() -> finches_lsvd

# recover row and column singular vectors
get_rows(finches_lsvd)
get_cols(finches_lsvd)
# augment ordination with point names and main effect estimates
augment_ord(finches_lsvd)
# summarize artifical coordinates (inertia is undefined)
tidy(finches_lsvd)
\dontrun{
  confer_inertia(finches_lsvd, .5)
}

# biplot (inertia is not defined and cannot be conferred)
finches_lsvd \%>\%
  augment_ord() \%>\%
  ggbiplot(aes(label = .name), sec.axes = "cols", scale.factor = 100) +
  theme_biplot() +
  geom_rows_vector(alpha = .5, color = "darkred") +
  geom_rows_text_radiate(size = 3, color = "darkred") +
  geom_cols_label(size = 3, alpha = .5, color = "royalblue3",
                  subset = grep("Geospiza", rownames(finches))) +
  geom_cols_point(alpha = .5, color = "royalblue3",
                  subset = -grep("Geospiza", rownames(finches))) +
  ggtitle(
    "Logistic SVD of the Galapagos island finches",
    "Islands (finches) scaled to the primary (secondary) axes"
  ) +
  expand_limits(x = c(-80, 60))
# site-species data frame of Sanderson Galapagos finches data
data(finches, package = "cooccur")
class(finches)
finches[seq(6L), seq(6L)]

# logistic PCA, omitting ubiquitous finches
finches \%>\%
  dplyr::filter(dplyr::if_any(where(is.integer), ~ . == 0)) \%>\%
  t() \%>\%
  logisticPCA_ord() \%>\%
  as_tbl_ord() \%>\%
  print() -> finches_lpca

# biplot (inertia is not defined and cannot be conferred)
finches_lpca \%>\%
  augment_ord() \%>\%
  ggbiplot(aes(label = .name), sec.axes = "cols", scale.factor = 50) +
  theme_biplot() +
  geom_rows_vector(alpha = .5, color = "darkred") +
  geom_rows_text_radiate(size = 3, color = "darkred") +
  geom_cols_label(size = 3, alpha = .5, color = "royalblue3",
                  subset = grep("Geospiza", rownames(finches))) +
  geom_cols_point(alpha = .5, color = "royalblue3",
                  subset = -grep("Geospiza", rownames(finches))) +
  ggtitle(
    "Logistic PCA of the Galapagos island finches",
    "Islands (finches) scaled to the primary (secondary) axes"
  ) +
  expand_limits(x = c(-25, 35))
# site-species data frame of Sanderson Galapagos finches data
data(finches, package = "cooccur")
class(finches)
finches[seq(6L), seq(6L)]

# convex logistic PCA, omitting ubiquitous finches
finches \%>\%
  dplyr::filter(dplyr::if_any(where(is.integer), ~ . == 0)) \%>\%
  t() \%>\%
  convexLogisticPCA_ord() \%>\%
  as_tbl_ord() \%>\%
  print() -> finches_clpca

# biplot (inertia is not defined and cannot be conferred)
finches_clpca \%>\%
  augment_ord() \%>\%
  ggbiplot(aes(label = .name), sec.axes = "cols", scale.factor = 50) +
  theme_biplot() +
  geom_rows_vector(alpha = .5, color = "darkred") +
  geom_rows_text_radiate(size = 3, color = "darkred") +
  geom_cols_label(size = 3, alpha = .5, color = "royalblue3",
                  subset = grep("Geospiza", rownames(finches))) +
  geom_cols_point(alpha = .5, color = "royalblue3",
                  subset = -grep("Geospiza", rownames(finches))) +
  ggtitle(
    "Convex logistic PCA of the Galapagos island finches",
    "Islands (finches) scaled to the primary (secondary) axes"
  ) +
  expand_limits(x = c(-25, 35))
}
\references{
Landgraf AJ, Lee Y (2020) "Dimensionality reduction for binary data through
the projection of natural parameters". \emph{Journal of Multivariate Analysis}
\strong{180}: 104668. \doi{10.1016/j.jmva.2020.104668}

Landgraf AJ, Lee Y (2020) "Dimensionality reduction for binary data through
the projection of natural parameters". \emph{Journal of Multivariate Analysis}
\strong{180}: 104668. \doi{10.1016/j.jmva.2020.104668}

Landgraf AJ, Lee Y (2020) "Dimensionality reduction for binary data through
the projection of natural parameters". \emph{Journal of Multivariate Analysis}
\strong{180}: 104668. \doi{10.1016/j.jmva.2020.104668}
}
\seealso{
Other methods for singular value decomposition-based techniques: 
\code{\link{methods-ade4}},
\code{\link{methods-ca-ca}},
\code{\link{methods-ca-mjca}},
\code{\link{methods-candisc-cancor}},
\code{\link{methods-nipals}},
\code{\link{methods-pma-cca}},
\code{\link{methods-pma-spc}}

Other methods for singular value decomposition-based techniques: 
\code{\link{methods-ade4}},
\code{\link{methods-ca-ca}},
\code{\link{methods-ca-mjca}},
\code{\link{methods-candisc-cancor}},
\code{\link{methods-nipals}},
\code{\link{methods-pma-cca}},
\code{\link{methods-pma-spc}}

Other methods for singular value decomposition-based techniques: 
\code{\link{methods-ade4}},
\code{\link{methods-ca-ca}},
\code{\link{methods-ca-mjca}},
\code{\link{methods-candisc-cancor}},
\code{\link{methods-nipals}},
\code{\link{methods-pma-cca}},
\code{\link{methods-pma-spc}}
}
\concept{methods for singular value decomposition-based techniques}
\concept{models from the \strong{logisticPCA} package}
