% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordr-extra.r
\docType{package}
\name{ordr.extra}
\alias{ordr.extra}
\alias{ordr.extra-package}
\title{\strong{ordr.extra} package}
\description{
Additional accessors to represent ordination models as
'tbl_ord's in \strong{\link[ordr:ordr]{ordr}}.
}
\details{
\strong{ordr} includes many methods and steps for extended workflows for 'tbl_ord'
objects. Accessors are required to wrap ordination models as 'tbl_ord's and
make these methods available to them. To conserve size, \strong{ordr} only
includes accessors for the most commonly distributed and used ordination
models. This package serves as a reservoir of accessors for additional, less
common models.
}
\examples{
# CA of hair & eye color data
ordinate(apply(HairEyeColor, seq(2L), sum), ca::ca)

# MCA of hair & eye color data, stratified by sex
ordinate(HairEyeColor, ca::mjca)

# NIPALS of air quality data
ordinate(airquality, ~ ade4::nipals(.), cols = seq(4L))
ordinate(airquality, cols = c("Ozone", "Solar.R", "Wind", "Temp"), ade4::nipals)

# sparse CCA of savings data
ordinate(
  x = LifeCycleSavings[, c(2L, 3L)],
  z = LifeCycleSavings[, c(1L, 4L, 5L)],
  model = PMA::CCA, K = 2L, penaltyx = .5, penaltyz = .5,
  # prevent errors
  typex = "standard", typez = "standard"
)

# lSVD of finches data
data(finches, package = "cooccur")
ordinate(finches, logisticSVD_ord)
}
\author{
\strong{Maintainer}: Jason Cory Brunson \email{cornelioid@gmail.com} (\href{https://orcid.org/0000-0003-3126-9494}{ORCID})

}
