% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordr-extra.r
\docType{package}
\name{ordr.extra}
\alias{ordr.extra}
\alias{ordr.extra-package}
\title{\strong{ordr.extra} package}
\description{
Additional accessors to represent ordination models as
'tbl_ord's in \strong{\link[ordr:ordr]{ordr}}.
}
\details{
\strong{ordr} includes many methods and steps for extended workflows for 'tbl_ord'
objects. Accessors are required to wrap ordination models as 'tbl_ord's and
make these methods available to them. To conserve size, \strong{ordr} only
includes accessors for the most commonly distributed and used ordination
models. This package serves as a reservoir of accessors for additional, less
common models.
}
