% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordr-extra.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tbl_ord}
\alias{recover_rows}
\alias{recover_cols}
\alias{recover_inertia}
\alias{recover_coord}
\alias{recover_conference}
\alias{augmentation_rows}
\alias{augmentation_cols}
\alias{augmentation_coord}
\alias{supplementation_rows}
\alias{supplementation_cols}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ordr}{\code{\link[ordr:tbl_ord]{as_tbl_ord}}, \code{\link[ordr:augmentation]{augmentation_cols}}, \code{\link[ordr:augmentation]{augmentation_coord}}, \code{\link[ordr:augmentation]{augmentation_rows}}, \code{\link[ordr:accessors]{recover_cols}}, \code{\link[ordr:conference]{recover_conference}}, \code{\link[ordr:accessors]{recover_coord}}, \code{\link[ordr:accessors]{recover_inertia}}, \code{\link[ordr:accessors]{recover_rows}}, \code{\link[ordr:supplementation]{supplementation_cols}}, \code{\link[ordr:supplementation]{supplementation_rows}}}
}}

