% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordr-extra.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tbl_ord}
\alias{recover_rows}
\alias{recover_cols}
\alias{recover_inertia}
\alias{recover_coord}
\alias{recover_conference}
\alias{recover_aug_rows}
\alias{recover_aug_cols}
\alias{recover_aug_coord}
\alias{recover_supp_rows}
\alias{recover_supp_cols}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ordr}{\code{\link[ordr:tbl_ord]{as_tbl_ord}}, \code{\link[ordr:augmentation]{recover_aug_cols}}, \code{\link[ordr:augmentation]{recover_aug_coord}}, \code{\link[ordr:augmentation]{recover_aug_rows}}, \code{\link[ordr:recoverers]{recover_cols}}, \code{\link[ordr:conference]{recover_conference}}, \code{\link[ordr:recoverers]{recover_coord}}, \code{\link[ordr:recoverers]{recover_inertia}}, \code{\link[ordr:recoverers]{recover_rows}}, \code{\link[ordr:supplementation]{recover_supp_cols}}, \code{\link[ordr:supplementation]{recover_supp_rows}}}
}}

