% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-wrap.r
\name{wrap-ord-extra}
\alias{wrap-ord-extra}
\alias{logisticSVD_ord}
\alias{logisticPCA_ord}
\alias{convexLogisticPCA_ord}
\alias{nipals_ord}
\alias{empca_ord}
\title{Wrappers for lossy ordination methods}
\usage{
logisticSVD_ord(
  x,
  k = 2,
  quiet = TRUE,
  max_iters = 1000,
  conv_criteria = 1e-05,
  random_start = FALSE,
  ...,
  partial_decomp = TRUE,
  main_effects = TRUE
)

logisticPCA_ord(
  x,
  k = 2,
  m = 4,
  quiet = TRUE,
  partial_decomp = FALSE,
  max_iters = 1000,
  conv_criteria = 1e-05,
  random_start = FALSE,
  ...,
  main_effects = TRUE
)

convexLogisticPCA_ord(
  x,
  k = 2,
  m = 4,
  quiet = TRUE,
  partial_decomp = FALSE,
  max_iters = 1000,
  conv_criteria = 1e-06,
  random_start = FALSE,
  ...,
  main_effects = TRUE,
  ss_factor = 4
)

nipals_ord(
  x,
  ncomp = min(nrow(x), ncol(x)),
  center = TRUE,
  scale = TRUE,
  maxiter = 500,
  tol = 1e-06,
  startcol = 0,
  fitted = FALSE,
  force.na = FALSE,
  gramschmidt = TRUE,
  verbose = FALSE
)

empca_ord(
  x,
  w,
  ncomp = min(nrow(x), ncol(x)),
  center = TRUE,
  scale = TRUE,
  maxiter = 100,
  tol = 1e-06,
  seed = NULL,
  fitted = FALSE,
  gramschmidt = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{...}{Additional parameters passed to original functions.}
}
\value{
Objects slightly modified from the outputs of the original functions,
with new '*-ord' classes.
}
\description{
These \verb{*_ord} functions wrap functions imported from other
packages with modifications for use with '\link{tbl_ord}' methods. They
complement those provided by \strong{ordr} for some common functions. Some
parameters are hidden from the user and set to settings required for these
methods, some matrix outputs are given row or column names to be used by
them, and new '*_ord' S3 class attributes are added to enable them.
}
\details{
The following table summarizes the wrapped functions:\tabular{llll}{
   Original function \tab Hide params \tab Add names \tab New class \cr
   \code{\link[logisticPCA:logisticSVD]{logisticPCA::logisticSVD()}} \tab No \tab Yes \tab No \cr
   \code{\link[logisticPCA:logisticPCA]{logisticPCA::logisticPCA()}} \tab No \tab Yes \tab No \cr
   \code{\link[logisticPCA:convexLogisticPCA]{logisticPCA::convexLogisticPCA()}} \tab No \tab Yes \tab No \cr
   \code{\link[nipals:nipals]{nipals::nipals()}} \tab No \tab No \tab Yes \cr
   \code{\link[nipals:empca]{nipals::empca()}} \tab No \tab No \tab Yes \cr
}
}
\seealso{
\link[ordr:wrap-ord]{ordr::wrap-ord}
}
